requirements.txt
requests==2.25.1
pydantic==1.8.2
PyYAML==5.4.1
apache-airflow==1.10.15
hdfs==2.6.0
